[Refactor]RoutineTableElementコンポーネント実装
